+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Changed "Noclip" effect to retry if a challenge is active.
+
 ## [2.1.0] - 2024-08-31
 
 - Added "Give Yoyo" effect.